move plot template into plot.cpp
